refactor: アイコンの定義を src/icons.rs に集約し Glyph 型で表す
